+++ src/main/java/com/crawler/extractor/model/CrawlerConf.java
@@ -4,35 +4,28 @@ import java.util.Date;
 import org.bson.types.ObjectId;
 import org.springframework.data.annotation.Id;
 import org.springframework.data.mongodb.core.mapping.Document;
+import com.fasterxml.jackson.annotation.JsonProperty;
 
 /**
+ * Class that represents collection of Crawler_conf;
+ * 
  * @author Dmytro Bilyi
  *
  * @data 09 January 2018
  *
- *       <p>
- *       Class that represents collection of Crawler_conf;
  */
 @Document(collection = "crawler_conf")
 public class CrawlerConf {
 
 	@Id
 	private ObjectId id;
-	private int max_number_active_crawler;
+	@JsonProperty(value = "max_number_active_crawler")
+	private int maxNumberActiveCrawler;
 	private Date createdDate;
 	private Date modifiedDate;
 
 
 	public CrawlerConf() {}
-
-	public CrawlerConf(ObjectId id, int max_number_active_crawler, Date createdDate,
-			Date modifiedDate) {
-		this();
-		this.id = id;
-		this.max_number_active_crawler = max_number_active_crawler;
-		this.createdDate = createdDate;
-		this.modifiedDate = modifiedDate;
-	}
 
 	public ObjectId getId() {
 		return id;
@@ -43,11 +36,11 @@ public class CrawlerConf {
 	}
 
 	public int getMax_number_active_crawler() {
-		return max_number_active_crawler;
+		return maxNumberActiveCrawler;
 	}
 
-	public void setMax_number_active_crawler(int max_number_active_crawler) {
-		this.max_number_active_crawler = max_number_active_crawler;
+	public void setMax_number_active_crawler(int maxNumberActiveCrawler) {
+		this.maxNumberActiveCrawler = maxNumberActiveCrawler;
 	}
 
 	public Date getCreatedDate() {
